refactor: reorganize database schema files and update imports
