+++ drizzle.config.ts
@@ -3,7 +3,7 @@ import { type Config } from "drizzle-kit";
 import { env } from "~/env";
 
 export default {
-  schema: "./src/server/db/schema.ts",
+  schema: "./src/server/db/schemas/index.ts",
   dialect: "postgresql",
   dbCredentials: {
     url: env.POSTGRES_URL,
