+++ src/app/api/uploadthing/core.ts
@@ -2,7 +2,7 @@ import { auth } from "@clerk/nextjs/server";
 import { createUploadthing, type FileRouter } from "uploadthing/next";
 import { UploadThingError } from "uploadthing/server";
 import { db } from "~/server/db";
-import { ImagesTable } from "~/server/db/schema";
+import { ImagesTable } from "~/server/db/schemas";
  
 const f = createUploadthing();
