+++ src/server/db/schemas/index.ts
@@ -1,0 +1,7 @@
+export * from './tags';
+export * from './favorites';
+export * from './recipes';
+export * from './images';
+export * from './ingredients';
+export * from './recipesToTags';
+export * from './recipeIngredients';
