+++ src/server/db/tableCreator.ts
@@ -1,0 +1,3 @@
+import { pgTableCreator } from 'drizzle-orm/pg-core';
+
+export const createTable = pgTableCreator((name) => `hyper-recipes_${name}`);
